Set up logging to Cloudwatch for apps and Nginx
The apps will log in json format to stdout, which is in turn written to
uWSGI's stdout, which is then written to a log file by supervisord and
logrotated. uWSGI logs are writted to stderr.

Nginx access logs are written in json to stdout and redirected to file
by supervisord, which are logrotated.

Cloudwatch agent streams these logs to cloudwatch.