Make base image pull and reuse cache from registry during build
By default, docker will only use cached layers that were created
locally when running the image build. This means that each time
the base image is build on a machine that didn't build the previous
docker will run the image creation steps from scratch, which is
likely to result in new layer snapshots.

In order to avoid this, we're pulling the latest base image from
Docker Hub and then using it as cache source during the build.

We don't need to do this for base-api and base-frontend at the
moment since they're not creating layers that can differ a lot
between separate runs (ie they're not installing packages or
modifying lots of files in non-repeatable ways).