Redirect awslogs agent logs to a file
awslogs agent is quite noisy, especially when running the container
locally without valid AWS credentials.

Since most of the time we're not interested in the messages generated
by awslogs agent but we might need them to investigate any potential
issues we can redirect the agent output to a 5MB file with supervisor.

This file is not itself shipped to CloudWatch, since any problems with
the agent we'd be interested in investigating would probably prevent it
from shipping logs to CloudWatch in the first place.