Add 'VERSION' and 'CHANGELOG' for python 3
The existing `VERSION` and `CHANGELOG` are renamed to `PY2...` versions
and a new `VERSION` and `CHANGELOG` are created for python3.

This will allow us to build and version the base images for py2 and py3
separately. As long as we're careful, we can build and tag each version
and push to Dockerhub without conflict. See the following commit for the
Makefile too see how.