Add rules to Makefile to support py2 and py3
In order to successfully build and then tag the base images for
different python versions we need to make sure that the tagging is done
immediately after the image is built. This is because the base-api and
base-frontend builds use the base image with the `latest` tag. This is
set when the image is built.

The images are also tagged with a version which is fetched from the
appropriate version file.