Set uwsgi worker memory limit to 256MB
We've had some issues with application workers being terminated
by OOM after exceeding the cgroups memory limit allocated to the
PaaS instance.

Requests that require a lot of memory to process require the app
worker to allocate additional memory for the process, most of which
python2 interpreter is not releasing back to the system after the
request has finished.

This causes the overall memory usage to increase overtime to the point
where next request with a large memory requirement will break the cgroup
memory limit, causing the OOM to terminate one of the workers (from
what we've seen, the worker processing the request is most likely to
get terminated) resulting in a 502 response to the user.

Setting `reload-on-rss` makes uwsgi restart workers if their memory usage
exceeds the set value.

Unlike OOM, uwsgi will only terminate workers between requests. This
means that worker can still use more memory during the request (as long
as the overall memory usage doesn't exceed PaaS instance limit), but it
will get replaced with a new worker after the request.

This allows us to reduce the average memory usage of the PaaS instance,
making it less likely to trigger OOM on large requests.

Since we're using instances of different sizes for different apps the
limit is set to 256MB. This should hopefully allow `search-api` and
frontend apps to gracefully recycle workers before running into the
cgroup limit without applying to the majority of `api` requests.

This doesn't remove the problem of request memory requirements,
since restarting the worker after memory usage spikes will increase
latency and reduce the overall throughput.

This also doesn't completely solve the 502 problem, since if a number
of workers are simultaneously processing expensive requests it's still
possible for them to exceed the instance memory limit.