Store logs from uWSGI stderr
Due to an issue with uWSGI, when running the apps on python3, all
application logs are coming through on uWSGI's stderr, instead of stdout
like before. This means they are mixed with uWSGI request logs as well
as startup logs.

Unfortunately it appears to be very difficult to get uWSGI to behave the
same way it does for python2. We've decided to just ship the logs from
stderr instead of trying to make it work with stdout.

We've disable logging in the uWSGI config, which prevents uWSGI request
logs from being written which quietens the log file down alot, but we
still get the startup logs. These are shipped to CludWatch, but they are
filtered out before hitting Kibana which is acceptable.