Set awslogs muti_line_start_pattern to default
As the application logs are now mixed with the uWSGI master process logs
we can no longer use the date format pattern as the marker for a new
log. It results in some application logs being followed by uWSGI logs
which then aren't shipped to Kibana. aws logs agent doesn't recognise a
new log message until it sees the date pattern again.

As all JSON logs are single lines, we can just use the start of a new
line as the pattern. This is the default behaviour by the logs agent, so
we can just remove the pattern definition here.