Only push images with `:latest` tag if python 3
Although we don't ever actually use the `:latest` tag, except when
building the images, it makes more sense to always know which python
version the `:latest` tag on Dockerhub is using.

This update will only push an image if it's version number is not equal
to '1', ensuring that python 2 images don't get pushed.