Update awslogs application log format to include milliseconds
Adding milliseconds to application logs allows us to keep the events
order (for messages coming from the same application instance) when
the logs are streamed to Elasticsearch.

AWS logs agent doesn't have a time format code for milliseconds, but
Python's strptime matches them with `%f` microsecond pattern, so we
can use it in our `datetime_format` to capture event times.

Unfortunately nginx doesn't have a time format that both includes
sub-second resolution and can be parsed by awslogs. Two options that
nginx provides are $ime_local (which doesn't include milliseconds)
or $msec, which is a UNIX timestamp and can't be parsed by strptime.